NSLoader.dll upgrades to newer version. Fixed minor impulse bug.

+++ Assets/NullSpace SDK/Demos/Scripts/Suit Demos/SuitImpulseDemo.cs
@@ -94,7 +94,7 @@ namespace NullSpace.SDK.Demos
 			}
 			set
 			{
-				effectDuration = (int)value;
+				effectDuration = value;
 			}
 		}
 		/// <summary>
